pyproject.toml
[project]
name = "webapp-training-project"
version = "0.1.0"
description = "Small Project to train webdev with uv"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.117.1",
    "flask>=3.1.2",
    "python-dotenv>=1.1.1",
]
